# Ensure Panama float vector distance impls inlinable 