--- a/lucene/CHANGES.txt
+++ b/lucene/CHANGES.txt
@@ -122,7 +122,7 @@ Optimizations
 * GITHUB#14032: Speed up PostingsEnum when positions are requested.
   (Adrien Grand)
 
-* GITHUB#14031: Ensure Panama float distance impls inlinable
+* GITHUB#14031: Ensure Panama float vector distance impls inlinable.
   (Robert Muir, Chris Hegarty)
 
 Bug Fixes