This commit reduces the Panama vector distance float implementations to less than the maximum bytecode size of a hot method to be inlined (325).

E.g. Previously: ` org.apache.lucene.internal.vectorization.PanamaVectorUtilSupport::dotProductBody (355 bytes)   failed to inline: callee is too large`.

After: `org.apache.lucene.internal.vectorization.PanamaVectorUtilSupport::dotProductBody (311 bytes)   inline (hot)`

This helps things a little.