@ChrisHegarty 2024-12-02
@rmuir nice!!! wanna push that to the branch? Then I’ll do some more benchmark runs tomorrow too.